+++ src/compiler/compile/Component.ts
@@ -10,6 +10,7 @@ import {
 	Scope,
 	extract_identifiers
 } from './utils/scope';
+import { Node as PeriscopicNode } from 'periscopic';
 import Stylesheet from './css/Stylesheet';
 import { test } from '../config';
 import Fragment from './nodes/Fragment';
@@ -183,9 +184,12 @@ export default class Component {
 		this.stylesheet.warn_on_unused_selectors(this);
 	}
 
-	add_var(variable: Var) {
+	add_var(variable: Var, add_to_lookup = true) {
 		this.vars.push(variable);
-		this.var_lookup.set(variable.name, variable);
+
+		if (add_to_lookup) {
+			this.var_lookup.set(variable.name, variable);
+		}
 	}
 
 	add_reference(name: string) {
@@ -216,6 +220,10 @@ export default class Component {
 				variable.subscribable = true;
 			}
 		} else {
+			if (this.compile_options.varsReport === 'full') {
+				this.add_var({ name, referenced: true }, false);
+			}
+
 			this.used_names.add(name);
 		}
 	}
@@ -340,19 +348,7 @@ export default class Component {
 			css,
 			ast: this.original_ast,
 			warnings: this.warnings,
-			vars: this.vars
-				.filter(v => !v.global && !v.internal)
-				.map(v => ({
-					name: v.name,
-					export_name: v.export_name || null,
-					injected: v.injected || false,
-					module: v.module || false,
-					mutated: v.mutated || false,
-					reassigned: v.reassigned || false,
-					referenced: v.referenced || false,
-					writable: v.writable || false,
-					referenced_from_script: v.referenced_from_script || false
-				})),
+			vars: this.get_vars_report(),
 			stats: this.stats.render()
 		};
 	}
@@ -400,6 +396,28 @@ export default class Component {
 				name: alias
 			};
 		};
+	}
+
+	get_vars_report(): Var[] {
+		const { compile_options, vars } = this;
+
+		const vars_report = compile_options.varsReport === false
+			? []
+			: compile_options.varsReport === 'full'
+				? vars
+				: vars.filter(v => !v.global && !v.internal);
+
+		return vars_report.map(v => ({
+			name: v.name,
+			export_name: v.export_name || null,
+			injected: v.injected || false,
+			module: v.module || false,
+			mutated: v.mutated || false,
+			reassigned: v.reassigned || false,
+			referenced: v.referenced || false,
+			writable: v.writable || false,
+			referenced_from_script: v.referenced_from_script || false
+		}));
 	}
 
 	error(
@@ -812,7 +830,7 @@ export default class Component {
 
 				if (node.type === 'AssignmentExpression' || node.type === 'UpdateExpression') {
 					const assignee = node.type === 'AssignmentExpression' ? node.left : node.argument;
-					const names = extract_names(assignee);
+					const names = extract_names(assignee as PeriscopicNode);
 
 					const deep = assignee.type === 'MemberExpression';